refactor(profile): move avatar editing to EditProfileScreen
