+++ src/screens/profile/EditProfileScreen.tsx
@@ -13,6 +13,10 @@ import { useStore } from '../../store/StoreProvider';
 import { showSuccessToast } from '../../utils/toast';
 import { getDisplayNameRules, getPhoneNumberRules } from '../../validation/rules';
 import { sg } from '../../styling';
+import EditableAvatar from '../../components/EditableAvatar';
+import * as ImagePicker from 'expo-image-picker';
+import Toast from 'react-native-toast-message';
+import { DEFAULT_AVATAR } from '../../constants/images';
 
 
 type FormValues = {
@@ -39,6 +43,30 @@ const EditProfileScreen = observer(() => {
     },
   });
 
+  const avatarSource = userProfile?.photoBase64
+  ? { uri: userProfile.photoBase64 }
+  : { uri: DEFAULT_AVATAR };
+
+  const handleChoosePhoto = async () => {
+    const result = await ImagePicker.launchImageLibraryAsync({
+      mediaTypes: ['images'],
+      allowsEditing: true,
+      quality: 0.8,
+    });
+
+    if (!result.canceled && result.assets?.[0]?.uri) {
+      try {
+        await authStore.uploadProfilePhoto(result.assets[0].uri, t);
+      } catch (err) {
+        Toast.show({
+          type: 'error',
+          text1: t('errors.error'),
+          text2: t('errors.uploadPhotoFailed'),
+        });
+      }
+    }
+  };
+
   const onSubmit = async (data: FormValues) => {
     await authStore.updateUserProfile(data);
 
@@ -54,6 +82,10 @@ const EditProfileScreen = observer(() => {
       <AppHeader showBack />
       <View style={styles.container}>
         <View>
+          <EditableAvatar
+            uri={avatarSource.uri}
+            onPress={handleChoosePhoto}
+          />
           <ControlledInput
           name="displayName"
           label={t('auth.displayName')}
@@ -85,7 +117,7 @@ const EditProfileScreen = observer(() => {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    paddingVertical: 24,
+    padding: 20,
     justifyContent:'space-between',
   },
 });
